+++ pilz_control/include/pilz_control/pilz_joint_trajectory_controller_impl.h
@@ -275,6 +275,16 @@ bool PilzJointTrajectoryController<SegmentImpl, HardwareInterface>::handleMonito
   return true;
 }
 
+template <class SegmentImpl, class HardwareInterface>
+void PilzJointTrajectoryController<SegmentImpl, HardwareInterface>::trajectoryCommandCB(
+    const JointTrajectoryConstPtr& /*msg*/)
+{
+  ROS_WARN_STREAM_NAMED(this->name_,
+                        "For safety reasons the trajectory command interface is deactivated"
+                            << " (for more information see https://github.com/ros-controls/ros_controllers/issues/493)."
+                            << " Please use the action interface instead.");
+}
+
 }  // namespace pilz_joint_trajectory_controller
 
 #endif  // PILZ_CONTROL_PILZ_JOINT_TRAJECTORY_CONTROLLER_IMPL_H