+++ pilz_control/test/unittest_pilz_joint_trajectory_controller.cpp
@@ -20,6 +20,10 @@
 #include <thread>
 
 #include <gtest/gtest.h>
+#include <gmock/gmock.h>
+
+#include <pilz_testutils/async_test.h>
+#include <pilz_testutils/ros_log_extender.h>
 
 #include <ros/ros.h>
 
@@ -45,6 +49,8 @@ static const std::string HOLD_SERVICE{ "/hold" };
 static const std::string UNHOLD_SERVICE{ "/unhold" };
 static const std::string IS_EXECUTING_SERVICE{ "/is_executing" };
 static const std::string TRAJECTORY_COMMAND_TOPIC{ "/command" };
+
+static const std::string LOG_MSG_RECEIVED_EVENT{ "log_msg_received" };
 
 using namespace pilz_joint_trajectory_controller;
 
@@ -386,7 +392,8 @@ static testing::AssertionResult InvokeIsExecutingServiceCallback(const Controlle
  * @brief For testing both the isExecuting method and the is_executing service callback we use parameterized tests.
  */
 class PilzJointTrajectoryControllerIsExecutingTest : public testing::Test,
-                                                     public testing::WithParamInterface<InvokeIsExecuting>
+                                                     public testing::WithParamInterface<InvokeIsExecuting>,
+                                                     public testing::AsyncTest
 {
 protected:
   void SetUp() override;
@@ -486,6 +493,11 @@ TEST_P(PilzJointTrajectoryControllerIsExecutingTest, testActionGoalExecution)
   EXPECT_FALSE(is_executing_result) << "There should be no execution to detect";
 }
 
+using pilz_testutils::IsWarn;
+using ::testing::_;
+/**
+ * @brief Tests that the controller does not execute if a trajectory is sent via command interface.
+ */
 TEST_P(PilzJointTrajectoryControllerIsExecutingTest, testTrajCommandExecution)
 {
   ASSERT_TRUE(performFullControllerStartup(&robot_driver_));
@@ -495,19 +507,21 @@ TEST_P(PilzJointTrajectoryControllerIsExecutingTest, testTrajCommandExecution)
       nh.advertise<trajectory_msgs::JointTrajectory>(CONTROLLER_NAMESPACE + TRAJECTORY_COMMAND_TOPIC, 1);
 
   GoalType goal{ generateSimpleGoal() };
+
+  pilz_testutils::ROSLogExtender ros_log_extender;
+  EXPECT_CALL(*ros_log_extender,
+              append(IsWarn("For safety reasons the trajectory command interface is deactivated "
+                            "(for more information see https://github.com/ros-controls/ros_controllers/issues/493). "
+                            "Please use the action interface instead."),
+                     _))
+      .WillOnce(ACTION_OPEN_BARRIER_VOID(LOG_MSG_RECEIVED_EVENT));
+
   trajectory_command_publisher.publish(goal.trajectory);
 
-  EXPECT_TRUE(updateUntilRobotMotion<RobotDriver>(&robot_driver_));
-
-  bool is_executing_result;
-  EXPECT_TRUE(invokeIsExecuting(is_executing_result));
-  EXPECT_TRUE(is_executing_result) << "Failed to detect trajectory command execution";
-
-  progressInTime(getGoalDuration(goal) + ros::Duration(DEFAULT_UPDATE_PERIOD_SEC));
-  robot_driver_.update();
-
-  EXPECT_TRUE(invokeIsExecuting(is_executing_result));
-  EXPECT_FALSE(is_executing_result) << "There should be no execution to detect";
+  ros::Duration observation_time = getGoalDuration(goal) + ros::Duration(DEFAULT_UPDATE_PERIOD_SEC);
+  EXPECT_TRUE(noExecutionObserving<RobotDriver>(&robot_driver_,
+                                                std::chrono::milliseconds(observation_time.toNSec() * 1000000ll)));
+  BARRIER(LOG_MSG_RECEIVED_EVENT);
 }
 
 TEST_P(PilzJointTrajectoryControllerIsExecutingTest, testStopTrajExecutionAtHold)